+++ php/commands/wp.php
@@ -5,6 +5,20 @@ use Terminus\Helpers\Input;
 use Terminus\Models\Collections\Sites;
 
 class WPCLI_Command extends CommandWithSSH {
+  /**
+   * Name of client that command will be run on server via
+   */
+  protected $client = 'WP-CLI';
+
+  /**
+   * A hash of commands which do not work in Terminus
+   * The key is the drush command
+   * The value is the Terminus equivalent, blank if DNE
+   */
+  protected $unavailable_commands = array(
+    'import' => '',
+    'db'     => '',
+  );
 
   /**
    * Invoke `wp` commands on a Pantheon development site
@@ -23,9 +37,11 @@ class WPCLI_Command extends CommandWithSSH {
    *
    */
   function __invoke( $args, $assoc_args ) {
+    $command = implode( $args, ' ' );
+    $this->checkCommand($command);
+    $sites       = new Sites();
+    $site        = $sites->get(Input::sitename($assoc_args));
     $environment = Input::env($assoc_args);
-    $sites = new Sites();
-    $site = $sites->get(Input::sitename($assoc_args));
     if (!$site) {
       $this->failure('Command could not be completed. Unknown site specified.');
     }
@@ -49,7 +65,6 @@ class WPCLI_Command extends CommandWithSSH {
       unset($assoc_args['env']);
     }
     # Create user-friendly output
-    $command = implode( $args, ' ' );
     $flags = '';
     foreach ( $assoc_args as $k => $v ) {
       if (isset($v) && (string) $v != '') {